pyproject.toml
[project]
name = "wisey-telemetry"
version = "0.1.0"
description = "Reusable OpenTelemetry setup for Wewise microservices"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "fastapi",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-exporter-jaeger-thrift",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-exporter-prometheus",
    "prometheus-client"
]

[tool.setuptools]
packages = ["wisey_telemetry"]
package-dir = {"" = "src"}
